Create README.md
# ERC20SilverTokenCreation
This is a simple simulated token creation on the ERC20 blockchain.
There are a few things that must be considered when creating a token and they are: 
a. Total supply of the token
b. Name of the token
c. Symbol of the token.
d. Balance of the token.
e. Capability to transfer the token.
In this Token I equally approved someone to spend the token